% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall.R
\name{build_density}
\alias{build_density}
\title{A simple density plot for the waterfall function}
\usage{
build_density(x, x.lims, fill, col = "black", bw = "SJ",
  show.underline = T, alpha = 0.8, label.expand = 0)
}
\arguments{
\item{x}{The data to be displayed.}

\item{x.lims}{The range of values which should be displayed.}

\item{fill}{The color which will be used for the inside plot area.
This value is passed to \code{\link[ggplot2]{geom_density}}.}

\item{col}{The color which will be used for the plot border.
This value is passed to \code{\link[ggplot2]{geom_density}}.}

\item{bw}{The smoothing bandwidth to be used.
This value is passed to \code{\link[ggplot2]{geom_density}}.
The default value is \code{\link[stats]{bw.SJ}}.}

\item{show.underline}{When \code{TRUE}, a horiznotal line will be drawn under
the plot.
The default value is \code{TRUE}.}

\item{alpha}{The alpha level for the inside area of the plot.
The default value is \code{0.8}.}

\item{label.expand}{The relative expand value for the right side.
The default value is \code{0.0}.}
}
\value{
ggplot-based plot which can be passed to the
  \code{\link{waterfall}} function.
}
\description{
A simple density plot for the waterfall function
}
