% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall.R
\name{build_ft}
\alias{build_ft}
\title{A simple frequency trail plot for the waterfall function}
\usage{
build_ft(x, x.lims, fill, col = "black", bw = "SJ",
  show.underline = T, alpha = 0.8, label.expand = 0)
}
\arguments{
\item{x}{The data to be displayed.}

\item{x.lims}{The range of values which should be displayed.}

\item{fill}{The color which will be used for the inside plot area.}

\item{col}{The color which will be used for the plot border.}

\item{bw}{The smoothing bandwidth to be used.
The value is passed to \code{\link[stats]{density}}.
The default value is \code{\link[stats]{bw.SJ}}.}

\item{show.underline}{When \code{TRUE}, a horiznotal line will be drawn under
the plot.
The default value is \code{TRUE}.}

\item{alpha}{The alpha level of individual plots.
The default value is \code{0.8}.}

\item{label.expand}{The relative expand value for the right side.
The default value is \code{0.0}.}
}
\description{
You can find the definition of frequency trails in
\href{http://www.brendangregg.com/FrequencyTrails/intro.html}{Brendan Gregg's Blog}
}
