% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatWaterfall.R
\name{compute_panel_waterfall}
\alias{compute_panel_waterfall}
\title{Title}
\usage{
compute_panel_waterfall(data, scales, width = 0.9)
}
\arguments{
\item{data}{}

\item{scales}{}

\item{width}{}
}
\description{
Title
}
