% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_waterfall.R
\name{geom_waterfall_label}
\alias{geom_waterfall_label}
\title{Title}
\usage{
geom_waterfall_label(..., lineheight = 0.8)
}
\arguments{
\item{...}{}

\item{mapping}{}

\item{data}{}

\item{position}{}

\item{na.rm}{}

\item{show.legend}{}

\item{inherit.aes}{}
}
\description{
Title
}
