% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggwaterfall.R
\docType{package}
\name{ggwaterfall-package}
\alias{ggwaterfall}
\alias{ggwaterfall-package}
\title{R package for creating waterfall chars with ggplot2}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Václav Hausenblas \email{vaclav.hausenblas@gmail.com}

}
