% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{rdistr}
\alias{rdistr}
\title{Random distribution}
\usage{
rdistr(n = 100, min = 0, max = 100)
}
\arguments{
\item{n}{The number of observations.}

\item{min}{The recommended minimum value.}

\item{max}{The recommended maximum value.}
}
\description{
Random distribution
}
