% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_waterfall.R
\name{stat_waterfall}
\alias{stat_waterfall}
\title{Title}
\usage{
stat_waterfall(
  geom = ggplot2::GeomRect,
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{geom}{}

\item{mapping}{}

\item{data}{}

\item{position}{}

\item{na.rm}{}

\item{show.legend}{}

\item{inherit.aes}{}

\item{...}{}
}
\description{
Title
}
