% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall.R
\name{waterfall}
\alias{waterfall}
\title{Generic waterfall plot}
\usage{
waterfall(x, build_plot, palette = NULL, bg = "white", col = "black",
  metric = NULL, calc.label = NULL, sort.value = NULL,
  show.labels = F, show.axis = T, show.underline = T,
  hide.outliers = F, title = NULL, bw = "SJ", overlap = 0.4,
  alpha = 0.8, label.expand = 0.15, style.pulsar = F)
}
\arguments{
\item{x}{List of vectors which will be displayed.
Each list element will be transformed to an individual plot.}

\item{build_plot}{A function which draws a plot for each vector.
This function should have the following signature:
`function(x, x.lims, fill, col, bw, show.underline, alpha)`.
There are two predefined functions which can be used:
\code{\link{build_density}} and \code{\link{build_ft}}.}

\item{palette}{Palette which will be used for plots.
Some palette examples: \itemize{
  \item \code{colorRampPalette(c("red", "orange"))(8)},
  \item \code{colorRampPalette(c("yellow", "yellow4"))(9)},
  \item \code{colorRampPalette(c("royalblue", "royalblue4"))(10)},
  \item \code{colorRampPalette(c("purple", "purple4"))(11)},
  \item \code{colorRampPalette(c("green2", "green4"))(12)}
  \item \code{rainbow(7)}
}
The default value is
\code{colorRampPalette(c("red", "orange"))(min(10, length(x)))}.}

\item{bg}{The background color.
The default value is \code{"white"}.}

\item{col}{The line color which will be used for axis lines, text labels,
and plot contours.
The default value is \code{"black"}.}

\item{metric}{An optional function which returns a single numeric value for
a numeric vector. Can be used in labels (when \code{show.labels} is
\code{TRUE}, the metric value van be used in \code{calc.label}) and
for sorting order.}

\item{calc.label}{A function which calculates labels for each plot; should be
used with \code{show.labels = TRUE}. The function should have the following
signature: \code{function(index, name, metric.value)}.
If \code{calc.label} is \code{NULL}, the following default algorithm will
be applied: if an \code{x} element has a name label, this name will be
used; otherwise if \code{metric} is defined, the metric value will be used;
otherwise the list index will be used.
The default value is \code{NULL}.}

\item{sort.value}{Defines how the individual plots will be sorted.
If \code{sort.value} is \code{NULL}, the plots will not be sorted.
If \code{sort.value} is \code{"metric"} and \code{metric} is defined,
the plots will be sorted by the given metric function.
Otherwise, \code{sort.value} should be a function which takes a numeric
vector and returns a single numeric value which will be used for sorting.
The default value is \code{NULL}.}

\item{show.labels}{When \code{TRUE}, labels for each individual plot will be
drawn on the right side. The text of the labels is contrlled by
the \code{calc.labels} function.
The default value is \code{FALSE}.}

\item{show.axis}{When \code{TRUE}, a horizontal axis with ticks will be drawn
at the bottom of the whole waterfall plot.
The default value is \code{TRUE}.}

\item{show.underline}{When \code{TRUE}, a horiznotal line will be drawn under
each individual plot.
The default value is \code{TRUE}.}

\item{hide.outliers}{When \code{TRUE}, the outlier values will be
automatically removed from each \code{x} element based on the
\href{https://en.wikipedia.org/wiki/Outlier#Tukey's_fences}{Tukey's fences}.
The default value is \code{FALSE}.}

\item{title}{The text for the title.}

\item{bw}{The smoothing bandwidth to be used. Will be passed to
\code{build_plot}.
The default value is \code{\link[stats]{bw.SJ}}.}

\item{overlap}{The overlap factor of the individual plots.
If \code{overlap} is \code{0}, the plots will not be overlapped.
If \code{overlap} is \code{1}, all the plots will be drawn on the same place.
The default value is \code{0.4}.}

\item{alpha}{The alpha level for the inside area of the each individual plot.
The default value is \code{0.8}.}

\item{label.expand}{The relative expand value for the right side.
Will be used only if \code{show.labels} is \code{TRUE}.
The default value is \code{0.15}.}

\item{style.pulsar}{Applies the plot style which is similar to
the iconic plot of a pulsar and the
\href{https://en.wikipedia.org/wiki/Unknown_Pleasures}{"Unknown Pleasures"}
Joy Division album cover. If it's true, the values of \code{palette},
\code{bg}, \code{col}, \code{alpha}, \code{show.axis},
\code{show.underline} will be ignored.
The default value is \code{FALSE}.}
}
\description{
Draw a composition of overlapped plots.
}
\details{
Inspired by the series of
\href{http://www.brendangregg.com/frequencytrails.html}{blog posts}
by Brendan Gregg.
}
\examples{
waterfall(rdistrs(5), build_plot = build_density)
waterfall(rdistrs(5), build_plot = build_ft)

}
