% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall.R
\docType{data}
\name{waterfall_colors}
\alias{waterfall_colors}
\title{color palette}
\format{An object of class \code{character} of length 3.}
\usage{
waterfall_colors
}
\description{
color palette
}
\keyword{datasets}
