% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall.R
\name{waterfall}
\alias{waterfall}
\title{Compute waterfall data}
\usage{
waterfall(data, detail_var = NULL, base_var = NULL, by_var = NULL,
  value_var = NULL, base_prefix = "", ordering = NULL)
}
\arguments{
\item{data}{}

\item{ordering}{}
}
\description{
Compute waterfall data
}
