% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall.R
\name{plot.waterfall}
\alias{plot.waterfall}
\title{Plot waterfall data}
\usage{
\method{plot}{waterfall}(
  wf_data,
  select = TRUE,
  scales = "fixed",
  flip = FALSE,
  xlim = NULL,
  ylim = c(min(wf_data$start, na.rm = TRUE), max(wf_data$end, na.rm = TRUE) * 1.1),
  nrow = NULL,
  ncol = NULL,
  label_font_size = 3,
  color_palette = NULL
)
}
\arguments{
\item{wf_data}{a data from waterfall() / waterfall_data()}

\item{select}{apply a filter}

\item{scales}{passed to ggplot2}

\item{flip}{flip chart to horizontal bars?}

\item{xlim}{passed to ggplot2}

\item{ylim}{passed to ggplot2}

\item{nrow}{passed to ggplot2}

\item{ncol}{passed to ggplot2}

\item{label_font_size}{passed to ggplot2}

\item{color_palette}{(optional) manual color_palette}
}
\description{
Plot waterfall data
}
