% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall.R
\name{waterfall}
\alias{waterfall}
\title{Compute waterfall data}
\usage{
waterfall(data, detail_var = NULL, base_var = NULL, by_var = NULL,
  value_var = NULL, drop_first = TRUE, base_prefix = "",
  ordering = NULL)
}
\arguments{
\item{data}{a data.frame object}

\item{detail_var}{name of the variable for item dimension}

\item{base_var}{name of the variable for totals}

\item{by_var}{name of the variable for facetting (optional)}

\item{value_var}{name of the variable with numeric values}

\item{drop_first}{drop observations for first base value (where differenced values are missing)}

\item{base_prefix}{prefix to be added to totals}

\item{ordering}{character vector with unique items (base_var) for manual ordering}
}
\description{
Compute waterfall data
}
