% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall.R
\name{plot.waterfall}
\alias{plot.waterfall}
\title{Plot waterfall data}
\usage{
\method{plot}{waterfall}(wf_data, select = TRUE, scales = "fixed",
  flip = FALSE, xlim = NULL, ylim = NULL, nrow = NULL,
  ncol = NULL, label_font_size = 3, color_palette = NULL)
}
\arguments{
\item{wf_data}{}

\item{color_palette}{}
}
\description{
Plot waterfall data
}
